Reallocated dev function to ignore it in coverage.
