+++ server/performanceUtils.ts
@@ -1,0 +1,40 @@
+import {
+  choiceKindType,
+  getChoiceDataFromPython,
+  getChoiceDataFromServer,
+} from "./pythonUtilities";
+
+// --------------------------------------------------------------------------
+// Temporary Function for Development
+// --------------------------------------------------------------------------
+export async function getChoiceListTimingTest(
+  enviroPath: string,
+  lineSoFar: string
+) {
+  // Compares the timing for choiceList using the server and vpython
+  // To use this, insert a call into getChoiceData()
+
+  let startTime: number = performance.now();
+  for (let index = 0; index < 10; index++) {
+    await getChoiceDataFromServer(
+      choiceKindType.choiceListTST,
+      enviroPath,
+      lineSoFar
+    );
+  }
+  let endTime: number = performance.now();
+  let deltaString: string = ((endTime - startTime) / 1000).toFixed(2);
+  console.log(`choiceList via the server 10x took: ${deltaString} seconds`);
+
+  startTime = performance.now();
+  for (let index = 0; index < 10; index++) {
+    getChoiceDataFromPython(
+      choiceKindType.choiceListTST,
+      enviroPath,
+      lineSoFar
+    );
+  }
+  endTime = performance.now();
+  deltaString = ((endTime - startTime) / 1000).toFixed(2);
+  console.log(`choiceList via vpython 10x took: ${deltaString} seconds`);
+}
