+++ server/pythonUtilities.ts
@@ -70,7 +70,7 @@ const emptyChoiceData: choiceDataType = {
   extraText: "",
 };
 
-async function getChoiceDataFromServer(
+export async function getChoiceDataFromServer(
   kind: choiceKindType,
   enviroPath: string,
   lineSoFar: string
@@ -169,39 +169,4 @@ export async function getHoverStringForRequirement(
     }
   }
   return returnValue;
-}
-
-// --------------------------------------------------------------------------
-// Temporary Function for Development
-// --------------------------------------------------------------------------
-export async function getChoiceListTimingTest(
-  enviroPath: string,
-  lineSoFar: string
-) {
-  // Compares the timing for choiceList using the server and vpython
-  // To use this, insert a call into getChoiceData()
-
-  let startTime: number = performance.now();
-  for (let index = 0; index < 10; index++) {
-    await getChoiceDataFromServer(
-      choiceKindType.choiceListTST,
-      enviroPath,
-      lineSoFar
-    );
-  }
-  let endTime: number = performance.now();
-  let deltaString: string = ((endTime - startTime) / 1000).toFixed(2);
-  console.log(`choiceList via the server 10x took: ${deltaString} seconds`);
-
-  startTime = performance.now();
-  for (let index = 0; index < 10; index++) {
-    getChoiceDataFromPython(
-      choiceKindType.choiceListTST,
-      enviroPath,
-      lineSoFar
-    );
-  }
-  endTime = performance.now();
-  deltaString = ((endTime - startTime) / 1000).toFixed(2);
-  console.log(`choiceList via vpython 10x took: ${deltaString} seconds`);
 }
